core: don't suppress host route to gateway in ip-config caputure
Why would we do this? The route is there, so, add it.
This revises commit 4fba2260f3e14a13f37499bd46ffcb220fdc5a4c
which added this check for matching generated connections.
I don't think this is still necessary, and if it is, then
the matching should be relaxed instead. It's bad to hide
routes from NMIP4Config/NMIP6Config, because those routes are
also exported via D-Bus.